Закончил второй экран
Добавил невидимый tempTemplate
Заменил удаление магазина статусом неактивности
Переделал header у drawer-а, добавил в источники шаблоны и историю
У Note появилось поле productId

+++ app/src/main/java/com/ultra/shopperlights2/Fragments/Fragment_Yellow_Purchase.java
@@ -11,6 +11,7 @@ import android.support.v4.app.FragmentTransaction;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -25,8 +26,10 @@ import com.ultra.shopperlights2.R;
 import com.ultra.shopperlights2.Units.*;
 import com.ultra.shopperlights2.Utils.Calc;
 import com.ultra.shopperlights2.Utils.ConfirmDialog;
+import com.ultra.shopperlights2.Utils.DateUtil;
 import com.ultra.shopperlights2.Utils.O;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 
@@ -41,13 +44,16 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 	 private ChangeYellowFragmentCallback callback;
 	 private YellowPurchaseListAdapter adapter;
 	 private Drawer drawer;
-	 private Spinner groupInput;
-	 private TextView statusTxt,totalPriceTxt;
+	 private Spinner inputSrcType,inputSrc;
+	 private Button btnAddNote;
+	 private TextView statusTxt,totalPriceTxt,srcTxt;
 	 private RecyclerView recyclerList;
 	 private long purchaseId;
-	 private HashMap<String,Boolean> usedMap;
+	 private HashMap<String,Boolean> mapUsed, mapEthereal;
 	 private BroadcastReceiver receiver;
-	 private static String lastTimeGroup;
+	 private static int lastTimeChoise,lastTimeSrcType= O.interaction.SRC_TYPE_CODE_GROUPS;
+	 private String srcTypes[]= {"группы","шаблоны","история"};
+	 private Template tempTemplate;
 
 	 private class Receiver extends BroadcastReceiver
 		 {
@@ -109,22 +115,50 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 			 note.setLocked(true);
 			 App.session.getNoteDao().update(note);
 			 adapter.addElement(note);
-			 usedMap.put(note.getTitle(),true);
+			 if(!note.isEthereal() )
+				 mapUsed.put(note.getTitle(),true);
+			 else
+				 mapEthereal.put(note.getTitle(),true);
 			 drawer.removeItem(drawerItemId);
 			 drawer.getDrawerLayout().setDrawerLockMode(DrawerLayout.LOCK_MODE_UNLOCKED);
-			 List<Product> productsAll= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId) ).list();
-			 List<Product> productsCompleted= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId),
-					ProductDao.Properties.Complete.eq(true) ).list();
-			 statusTxt.setText(productsCompleted.size() +"/"+ productsAll.size() );
+			 initStatusTxt();
 			 return true;
 			 }
 		 }
-	 private class GroupSelectListener implements AdapterView.OnItemSelectedListener
+	 private class SrcTypeSelectListener implements AdapterView.OnItemSelectedListener
 		 {
 		 @Override
 		 public void onItemSelected(AdapterView<?> parent,View view,int position,long id)
 			 {
-			 selectGroupSpinner();
+			 lastTimeSrcType=position;
+			 lastTimeChoise=0;
+			 switch(lastTimeSrcType)
+				 {
+				 case O.interaction.SRC_TYPE_CODE_GROUPS:
+					 srcTxt.setText("группы");
+					 btnAddNote.setVisibility(View.VISIBLE);
+					 break;
+				 case O.interaction.SRC_TYPE_CODE_TEMPLATES:
+					 srcTxt.setText("шаблоны");
+					 btnAddNote.setVisibility(View.INVISIBLE);
+					 break;
+				 case O.interaction.SRC_TYPE_CODE_HISTORY:
+					 srcTxt.setText("история");
+					 btnAddNote.setVisibility(View.INVISIBLE);
+					 break;
+				 }
+			 initSrcSpinner();
+			 }
+		 @Override
+		 public void onNothingSelected(AdapterView<?> parent) {}
+		 }
+	 private class SrcSelectListener implements AdapterView.OnItemSelectedListener
+		 {
+		 @Override
+		 public void onItemSelected(AdapterView<?> parent,View view,int position,long id)
+			 {
+			 lastTimeChoise=position;
+			 selectSrcSpinner();
 			 }
 		 @Override
 		 public void onNothingSelected(AdapterView<?> parent) {}
@@ -166,6 +200,8 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 			 session.getPurchaseDao().update(purchase);
 			 for(Note note : session.getNoteDao().queryBuilder().where(NoteDao.Properties.Locked.eq(true) ).list() )
 				 {
+				 if(mapEthereal.containsKey(note.getTitle() ) )
+					 continue;
 				 List<TagToNote> tagToNotes= session.getTagToNoteDao().queryBuilder().where(TagToNoteDao.Properties.NoteId.eq(note.getId() ) ).list();
 				 for(TagToNote tagToNote : tagToNotes)
 					 session.getTagToNoteDao().delete(tagToNote);
@@ -180,6 +216,7 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 			 }
 		 finally
 			 {
+			 clearTempTemplate();
 			 callback.changeYellowFragment(false);
 			 drawer.getDrawerLayout().setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
 			 }
@@ -198,8 +235,15 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 			 note.setLocked(false);
 			 session.getNoteDao().update(note);
 			 }
+		 clearTempTemplate();
 		 callback.changeYellowFragment(false);
 		 drawer.getDrawerLayout().setDrawerLockMode(DrawerLayout.LOCK_MODE_LOCKED_CLOSED);
+		 }
+	 private void clearTempTemplate()
+		 {
+		 for(Note note : tempTemplate.getNotes())
+			 App.session.getNoteDao().delete(note);
+		 tempTemplate.getNotes().clear();
 		 }
 	 @Override
 	 public void delElement(String title)
@@ -207,7 +251,10 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 Note note= App.session.getNoteDao().queryBuilder().where(NoteDao.Properties.Title.eq(title),NoteDao.Properties.Locked.eq(true) ).list().get(0);
 		 note.setLocked(false);
 		 App.session.getNoteDao().update(note);
-		 usedMap.put(title,false);
+		 if(!note.isEthereal() )
+			 mapUsed.put(title,false);
+		 else
+			 mapEthereal.put(title,false);
 		 }
 	 @Override
 	 public void initDialog(long id)
@@ -217,28 +264,55 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 FragmentTransaction transaction= getFragmentManager().beginTransaction();
 		 dialog.show(transaction,"");
 		 }
+	 private ArrayList<String> getSources()
+		 {
+		 ArrayList<String> result= new ArrayList<>();
+		 switch(lastTimeSrcType)
+			 {
+			 case O.interaction.SRC_TYPE_CODE_GROUPS:
+				 result.add("");
+				 for(Group group : App.session.getGroupDao().queryBuilder().orderAsc(GroupDao.Properties.Priority).list())
+					 result.add(group.getTitle() );
+				 break;
+			 case O.interaction.SRC_TYPE_CODE_TEMPLATES:
+				 for(Template template : App.session.getTemplateDao().loadAll() )
+					 {
+					 if(template.getTitle().equals(O.TEMP_TEMPLATE_NAME) )
+						 continue;
+					 result.add(template.getTitle() );
+					 }
+				 break;
+			 case O.interaction.SRC_TYPE_CODE_HISTORY:
+				 List<Purchase> purchases= App.session.getPurchaseDao().queryBuilder().orderDesc(PurchaseDao.Properties.Date).limit(50).offset(1).list();
+				 int i=0;
+				 for(Purchase purchase : purchases)
+					 {
+					 Log.d(O.TAG,"getSources: purchase.id="+ purchase.getId() );
+					 if(purchase.getProducts().size() == 0)
+						 continue;
+					 long id=purchase.getId();
+					 String shop;
+					 long shopId=purchase.getShopId();
+					 if(shopId!=0)
+						 shop=App.session.getShopDao().load(shopId).getTitle();
+					 else
+						 shop="Unknown";
+					 String date=DateUtil.getDateStr(purchase.getDate() );
+					 result.add(id +" "+ shop +" ("+ date +")");
+					 i++;
+					 if(i==10)
+						 break;
+					 }
+				 break;
+			 }
+		 return result;
+		 }
 	 public void updateLists()
 		 {
-		 initMap();
-		 ArrayList<String> groups= new ArrayList<>();
-		 groups.add("");
-		 DaoSession session= App.session;
-		 for(Group group : session.getGroupDao().loadAll() )
-			 groups.add(group.getTitle() );
-		 ArrayAdapter<String> spinnerAdapter= new ArrayAdapter<>(getContext(),android.R.layout.simple_spinner_item,groups);
-		 spinnerAdapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-		 if(groupInput!=null)
-			{
-			 groupInput.setAdapter(spinnerAdapter);
-			 if(groups.contains(lastTimeGroup) )
-				 groupInput.setSelection(groups.indexOf(lastTimeGroup) );
-			 }
 		 initAdapter();
-		 selectGroupSpinner();
-		 List<Product> productsAll= session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId) ).list();
-		 List<Product> productsCompleted= session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId),
+		 initStatusTxt();
+		 List<Product> productsCompleted= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId),
 				 ProductDao.Properties.Complete.eq(true) ).list();
-		 statusTxt.setText(productsCompleted.size() +"/"+ productsAll.size() );
 		 float totalPrice=0;
 		 for(Product product : productsCompleted)
 			 totalPrice+= product.getPrice() * (product.getN()==0 ? 1 : product.getN() );
@@ -249,21 +323,76 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 callback=_callback;
 		 drawer=_drawer;
 		 }
-	 private void selectGroupSpinner()
+	 private void initStatusTxt()
 		 {
-		 if(groupInput!=null)
+		 List<Product> productsAll= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId) ).list();
+		 List<Product> productsCompleted= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId),
+				 ProductDao.Properties.Complete.eq(true) ).list();
+		 statusTxt.setText(productsCompleted.size() +"/"+ productsAll.size() );
+		 }
+	 private void initSrcTypeSpinner()
+		 {
+		 if(inputSrcType!=null)
 			 {
-			 lastTimeGroup= groupInput.getSelectedItem().toString();
-			 if(lastTimeGroup.length()==0)
+			 ArrayAdapter<String> adapterSrcTypes= new ArrayAdapter<>(getContext(),android.R.layout.simple_spinner_item,srcTypes);
+			 adapterSrcTypes.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+			 inputSrcType.setAdapter(adapterSrcTypes);
+			 }
+		 }
+	 private void initSrcSpinner()
+		 {
+		 if(inputSrc!=null)
+			 {
+			 ArrayList<String> sources= getSources();
+			 ArrayAdapter<String> adapterSources= new ArrayAdapter<>(getContext(),android.R.layout.simple_spinner_item,sources);
+			 adapterSources.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+			 inputSrc.setSelection(0);
+			 inputSrc.setAdapter(adapterSources);
+			 }
+		 }
+	 private void selectSrcSpinner()
+		 {
+		 if(inputSrc!=null)
+			 {
+			 inputSrcType.setSelection(lastTimeSrcType);
+			 inputSrc.setSelection(lastTimeChoise);
+			 ArrayList<Note> notes= new ArrayList<>();
+			 String selectedItem=inputSrc.getSelectedItem().toString();
+			 switch(lastTimeSrcType)
 				 {
-				 ArrayList<Note> notes= new ArrayList<>(App.session.getNoteDao().queryBuilder().where(NoteDao.Properties.GroupId.eq(0) ).list() );
-				 initDrawerList(notes);
+				 case O.interaction.SRC_TYPE_CODE_GROUPS:
+					 if(selectedItem.length() == 0)
+						 notes.addAll(App.session.getNoteDao().queryBuilder().where(NoteDao.Properties.GroupId.eq(0),
+								 		NoteDao.Properties.Ethereal.eq(false) ).list() );
+					 else
+						 {
+						 Group group= App.session.getGroupDao().queryBuilder().where(GroupDao.Properties.Title.eq(selectedItem) )
+								 	.list().get(0);
+						 notes.addAll(group.getNotes() );
+						 }
+					 break;
+				 case O.interaction.SRC_TYPE_CODE_TEMPLATES:
+					 notes.addAll(App.session.getTemplateDao().queryBuilder().where(TemplateDao.Properties.Title.eq(selectedItem) )
+							 		.list().get(0).getNotes() );
+					 break;
+				 case O.interaction.SRC_TYPE_CODE_HISTORY:
+					 long selectedPurchaseId= Long.parseLong(selectedItem.split(" ")[0] );
+					 for(Product product : App.session.getPurchaseDao().load(selectedPurchaseId).getProducts() )
+						 {
+						 Note note= new Note();
+						 note.setTitle(product.getTitle() );
+						 note.setN(product.getN() );
+						 note.setEthereal(true);
+						 tempTemplate.getNotes().add(note);
+						 note.setTemplateId(tempTemplate.getId() );
+						 note.setProductId(product.getId() );
+						 App.session.getNoteDao().insert(note);
+						 Log.d(O.TAG,"selectSrcSpinner: notes="+ notes);
+						 notes.add(note);
+						 }
+					 break;
 				 }
-			 else
-				 {
-				 Group group= App.session.getGroupDao().queryBuilder().where(GroupDao.Properties.Title.eq(lastTimeGroup) ).list().get(0);
-				 initDrawerList(new ArrayList<>(group.getNotes() ) );
-				 }
+			 initDrawerList(notes);
 			 }
 		 }
 	 private void initDrawerList(ArrayList<Note> src)
@@ -271,7 +400,9 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 drawer.removeAllItems();
 		 for(Note note : src)
 			 {
-			 if(usedMap.get(note.getTitle() ) )
+			 if(mapUsed.containsKey(note.getTitle() ) && mapUsed.get(note.getTitle() ) )
+				 continue;
+			 if(mapEthereal.containsKey(note.getTitle() ) && mapEthereal.get(note.getTitle() ) )
 				 continue;
 			 PrimaryDrawerItem drawerItem= new PrimaryDrawerItem().withName(note.getTitle() );
 			 drawerItem.withTag(note.getId() );
@@ -283,19 +414,44 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 {
 		 purchaseId= App.session.getPurchaseDao().queryBuilder().where(PurchaseDao.Properties.Completed.eq(false) ).list().get(0).getId();
 		 }
-	 private void initMap()
+	 private void initMaps()
 		 {
-		 usedMap= new HashMap<>();
+		 mapUsed= new HashMap<>();
+		 mapEthereal= new HashMap<>();
 		 ArrayList<Note> notes=new ArrayList<>(App.session.getNoteDao().loadAll());
 		 for(Note note : notes)
-			 usedMap.put(note.getTitle(),false);
+			 if(!note.isEthereal() )
+				 mapUsed.put(note.getTitle(),false);
+			 else
+				 mapEthereal.put(note.getTitle(),false);
+		 }
+	 private void initTempTemplate()
+		 {
+		 List<Template> TTList=App.session.getTemplateDao().queryBuilder().where(TemplateDao.Properties.Title.eq(O.TEMP_TEMPLATE_NAME)).list();
+		 if(TTList.size()==0)
+			 {
+			 tempTemplate= new Template();
+			 tempTemplate.setTitle(O.TEMP_TEMPLATE_NAME);
+			 App.session.getTemplateDao().insert(tempTemplate);
+			 Log.d(O.TAG,"initAdapter: tempTemplate created");
+			 }
+		 else
+			 tempTemplate= TTList.get(0);
 		 }
 	 private void initAdapter()
 		 {
 		 List<Product> products= App.session.getProductDao().queryBuilder().where(ProductDao.Properties.PurchaseId.eq(purchaseId)).list();
 		 for(Product product : products)
-			 if(usedMap.containsKey(product.getTitle() ) )
-				 usedMap.put(product.getTitle(),true);
+			 if(!product.isEthereal() )
+				 {
+				 if(mapUsed.containsKey(product.getTitle() ) )
+					 mapUsed.put(product.getTitle(),true);
+				 }
+			 else
+				 {
+				 if(mapEthereal.containsKey(product.getTitle() ) )
+					 mapEthereal.put(product.getTitle(),true);
+				 }
 		 adapter= new YellowPurchaseListAdapter(getContext(),new ArrayList<>(products),purchaseId,O.actions.ACTION_FRAGMENT_YELLOW_PURCHASE,this,this);
 		 recyclerList.setAdapter(adapter);
 		 recyclerList.setLayoutManager(new LinearLayoutManager(getContext() ) );
@@ -308,6 +464,8 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 View mainView= inflater.inflate(R.layout.yellow_screen,container,false);
 
 		 getPurchaseId();
+		 initMaps();
+		 initTempTemplate();
 		 ImageButton buttonOpen= (ImageButton)mainView.findViewById(R.id.btn_openDrawer);
 		 recyclerList= (RecyclerView)mainView.findViewById(R.id.list);
 		 Button btnComplete= (Button)mainView.findViewById(R.id.completePurchase);
@@ -320,14 +478,19 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 			 drawer.getDrawerLayout().setDrawerLockMode(DrawerLayout.LOCK_MODE_UNLOCKED);
 			 View header= drawer.getHeader();
 			 ImageButton drawerClose= (ImageButton)header.findViewById(R.id.btn_close);
-			 Button btnAddNote= (Button)header.findViewById(R.id.btn_add);
-			 groupInput= (Spinner)header.findViewById(R.id.groupInput);
+			 btnAddNote= (Button)header.findViewById(R.id.btn_add);
+			 inputSrcType= (Spinner)header.findViewById(R.id.srcTypeInput);
+			 inputSrc= (Spinner)header.findViewById(R.id.srcInput);
+			 srcTxt= (TextView)header.findViewById(R.id.srcTxt);
 
 			 btnAddNote.setOnClickListener(new DrawerHeaderClickListener() );
 			 drawerClose.setOnClickListener(new DrawerHeaderClickListener() );
 			 drawer.setOnDrawerItemClickListener(new DrawerItemClickListener() );
-			 groupInput.setOnItemSelectedListener(new GroupSelectListener() );
-//			 updateLists();
+			 inputSrc.setOnItemSelectedListener(new SrcSelectListener() );
+			 inputSrcType.setOnItemSelectedListener(new SrcTypeSelectListener() );
+			 initSrcTypeSpinner();
+			 initSrcSpinner();
+			 selectSrcSpinner();
 			 }
 		 buttonOpen.setOnClickListener(new DrawerButtonClickListener() );
 		 btnComplete.setOnClickListener(new PurchaseButtonsListener() );
@@ -335,7 +498,6 @@ public class Fragment_Yellow_Purchase extends Fragment implements YellowScreenDe
 		 receiver= new Receiver();
 		 IntentFilter filter= new IntentFilter(O.actions.ACTION_FRAGMENT_YELLOW_PURCHASE);
 		 getContext().registerReceiver(receiver,filter);
-
 		 return mainView;
 		 }
 	 @Override
